requirements.txt
asn1crypto==0.22.0
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
Django==1.11.4
django-allauth==0.32.0
django-cleanup==1.0.1
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-auth==0.9.1
djangorestframework==3.6.3
enum34==1.1.6
Fabric==1.14.0
idna==2.5
ipaddress==1.0.18
oauthlib==2.0.2
olefile==0.44
paramiko==2.2.1
Pillow==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.3.1
pyasn1==0.3.3
pycparser==2.18
PyNaCl==1.1.2
python-openid==2.2.5
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.22
